requirements.txt
python-telegram-bot
ccxt
pandas
numpy